fix: close the floating window when window is split (`<C-w>v`)
Prior to this commit and usage of `<C-w>v` on the Hurl output float
window would case the output to be split into another window, but leave
the floating window behind covering the entire editor. At this point the
only way to close the floating window is to either target it
specifically with a window move command or to use your mouse to select
the window then close it out.