refactor: use vim.system instead of vim.fn.jobstart
With `vim.system` we get better type coersion and general support from
libuv. Jobstart is primarily meant to be used from vimscript, not lua.
Per `:h jobstart` it recommends using `vim.system`.

This also makes it easier to handle building the commands list for hurl
as now we only have to manage a table.

This is the first of a few commits related to #9 and #7.